backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
motor==3.3.2  # MongoDB async driver
pymongo==4.6.0
psycopg2-binary==2.9.9  # PostgreSQL (optional)

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.24.3
pandas==2.1.4
joblib==1.3.2

# Time Series Forecasting
prophet==1.1.4
statsmodels==0.14.0

# Deep Learning (Optional)
torch==2.1.1
torchaudio==2.1.1

# Speech and Language Processing
SpeechRecognition==3.10.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
gtts==2.4.0
langdetect==1.0.9
textblob==0.17.1

# Audio Processing
pyaudio==0.2.11
wave==0.0.2

# Emotion Analysis
transformers==4.35.2
vaderSentiment==3.3.2

# Weather and Location APIs
requests==2.31.0
geopy==2.4.0
python-dotenv==1.0.0

# Data Processing
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and Monitoring
loguru==0.7.2

# CORS and Security
python-cors==1.7.0

# Image Processing (for satellite data)
Pillow==10.1.0
opencv-python==4.8.1.78

# Geocoding
geocoder==1.38.1

# Async HTTP Client
aiohttp==3.9.1
aiofiles==23.2.0